Unified and improved decoding of arrays and strings
If assumed that a string in a S7-1200 has no special type-flags and is
transmitted as an array of USInt, then the dissection can be unified. If
so, the set type-flag bit 0x10 marks that an array with a size-header is
following. With this changes, the deconding has not to differ between
session and normal telegrams.

git-svn-id: svn://svn.code.sf.net/p/s7commwireshark/code/trunk@148 45802b70-cb78-474c-a225-0ca69f292a50