Added decoding of 2 bytes of the unknown last part of read/write request and response
The first byte of the last data set seems to be an ID:
It increases with each request by 1, starting from 0.
For responses it seems to be calculated from the request id.

The second byte of the last sequence seems to be the number of bytes which follow.
Except this value is 0.

Signed-off-by: Juergen Kosel <klj@softing.com>

git-svn-id: svn://svn.code.sf.net/p/s7commwireshark/code/trunk@215 45802b70-cb78-474c-a225-0ca69f292a50