Modifications to be compatible with Wireshark version 2.0
This is the minimal modification to let the plugin compile and
use with Wireshark 2.0 and later.
In Wireshark 2.0 the libwireshark api has changed.
The handy proto_tree_add_text function was removed. In first step
I've added a dummy-fake function which allows to use this function
for quick tests. Later all this calls should be replaced with
appropriate header fields.
As build system CMake is used instead of nmake.

git-svn-id: svn://svn.code.sf.net/p/s7commwireshark/code/trunk@285 45802b70-cb78-474c-a225-0ca69f292a50