Tagdescription: bitoffset for optimized and nonoptimized
If a DB is nonoptimized, both bitoffsets are the same.
If a DB is optimized, the variables are packed and sorted by their size.
The nonoptimized addresses can be used to get the order of variables as
they were configured in engineering tool.
Retain variables have their own memory area.

git-svn-id: svn://svn.code.sf.net/p/s7commwireshark/code/trunk@430 45802b70-cb78-474c-a225-0ca69f292a50